// ui/src/sdk/models/DtoSshRevealResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoGlue API
 * API for managing K3s clusters across cloud providers
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DtoSshRevealResponse
 */
export interface DtoSshRevealResponse {
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    created_at?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    fingerprint?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    organization_id?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    private_key?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    public_key?: string;
    /**
     * 
     * @type {string}
     * @memberof DtoSshRevealResponse
     */
    updated_at?: string;
}

/**
 * Check if a given object implements the DtoSshRevealResponse interface.
 */
export function instanceOfDtoSshRevealResponse(value: object): value is DtoSshRevealResponse {
    return true;
}

export function DtoSshRevealResponseFromJSON(json: any): DtoSshRevealResponse {
    return DtoSshRevealResponseFromJSONTyped(json, false);
}

export function DtoSshRevealResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DtoSshRevealResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'created_at': json['created_at'] == null ? undefined : json['created_at'],
        'fingerprint': json['fingerprint'] == null ? undefined : json['fingerprint'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'organization_id': json['organization_id'] == null ? undefined : json['organization_id'],
        'private_key': json['private_key'] == null ? undefined : json['private_key'],
        'public_key': json['public_key'] == null ? undefined : json['public_key'],
        'updated_at': json['updated_at'] == null ? undefined : json['updated_at'],
    };
}

export function DtoSshRevealResponseToJSON(json: any): DtoSshRevealResponse {
    return DtoSshRevealResponseToJSONTyped(json, false);
}

export function DtoSshRevealResponseToJSONTyped(value?: DtoSshRevealResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['created_at'],
        'fingerprint': value['fingerprint'],
        'id': value['id'],
        'name': value['name'],
        'organization_id': value['organization_id'],
        'private_key': value['private_key'],
        'public_key': value['public_key'],
        'updated_at': value['updated_at'],
    };
}